internal downloader: rename types.py -> manifest.py
