+++ ego4d/internal/download/manifest.py
